Made Ship attribute armor a double instead of int

+++ sailinggame/src/main/java/com/group11/model/Ship.java
@@ -8,7 +8,7 @@ import java.awt.Point;
 public class Ship extends AMovableBody implements HasWeapon {
 
     private int shipLevel;
-    private int armor;
+    private double armor;
     private int cannons;
     private boolean sailIsUp;
     private boolean anchorDown;
@@ -21,7 +21,7 @@ public class Ship extends AMovableBody implements HasWeapon {
      * @param cannons    - the cannons of the ship
      * @param hitPoints the number of hitpoints of the ship
      */
-    public Ship(Point pos, int shipLevel, int armor, int cannons, double hitPoints){
+    public Ship(Point pos, int shipLevel, double armor, int cannons, double hitPoints){
         super(pos, hitPoints, "A basic ship");
         this.shipLevel   = shipLevel;
         this.armor       = armor;
@@ -55,7 +55,7 @@ public class Ship extends AMovableBody implements HasWeapon {
      * Returns armor rating
      * @return armor rating
      */
-    public int getArmor() {
+    public double getArmor() {
         return this.armor;
     }
